syscall: handle spurious IRQ for preemption
Make the treatment of spurious IRQs equal between handleInterruptEntry
and preemption handling. The infoflow proof requires that these behave
identically -- this was so far only true for platforms that define
handleSpuriousIRQ to be empty (which is almost all of them).

Adding platforms to the verification that have a non-empty body for
handleSpuriousIRQ requires equalising the calls.

Factor out the code that is common between preemption and interrupt
kernel entry into its own function maybeHandleInterrupt to remove
duplication.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>